Restructure and additional edge case (positioning)
